+++ FitnessApp/Home/Models/ActivityModel.swift
@@ -1,0 +1,18 @@
+//
+//  ActivityModel.swift
+//  FitnessApp
+//
+//  Created by Nicolas Ströbel on 24.05.25.
+//
+
+import SwiftUI
+
+
+struct Activity {
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tintColor: Color
+    let amount: String
+}
