+++ FitnessApp/Home/Models/WorkoutModel.swift
@@ -1,0 +1,19 @@
+//
+//  WorkoutModel.swift
+//  FitnessApp
+//
+//  Created by Nicolas Ströbel on 24.05.25.
+//
+
+import SwiftUI
+
+
+struct Workout {
+    let id: Int
+    let title: String
+    let image: String
+    let tintColor: Color
+    let duration: String
+    let date: String
+    let calories: String
+}
