+++ FitnessApp/Home/Views/ActivityCardView.swift
@@ -8,17 +8,6 @@
 import SwiftUI
 
 
-
-struct Activity {
-    let id: Int
-    let title: String
-    let subtitle: String
-    let image: String
-    let tintColor: Color
-    let amount: String
-}
-
-
 struct ActivityCardView: View {
     
     @State var activity: Activity
