+++ FitnessApp/Home/Views/HomeView.swift
@@ -10,24 +10,9 @@ import SwiftUI
 
 struct HomeView: View {
     
-    @State var calories: Int = 123
-    @State var active: Int = 52
-    @State var stand: Int = 8
+    @StateObject var viewModel = HomeViewModel()
     
-    var mockActivities = [
-        Activity(id: 0, title: "Heute", subtitle: "Schritte", image: "figure.walk", tintColor: .blue, amount: "10,000"),
-        Activity(id: 1, title: "Morgen", subtitle: "Joggen", image: "figure.run", tintColor: .red, amount: "584"),
-        Activity(id: 2, title: "Gestern", subtitle: "Pausen", image: "figure.stand", tintColor: .orange, amount: "3"),
-        Activity(id: 3, title: "Übermorgen", subtitle: "Sex", image: "figure.socialdance.circle", tintColor: .green, amount: "5h")
-    ]
-    
-    var mockWorkouts = [
-        Workout(id: 0, title: "Running", image: "figure.run", tintColor: .green, duration: "23 mins", date: "Aug 8", calories: "312 kcal"),
-        Workout(id: 1, title: "Weight Lifting", image: "dumbbell", tintColor: .black, duration: "90 mins", date: "Aug 9", calories: "564 kcal"),
-        Workout(id: 2, title: "Basketball", image: "figure.basketball", tintColor: .orange, duration: "90 mins", date: "Aug 20", calories: "684 kcal"),
-        Workout(id: 3, title: "Fußball", image: "figure.soccer", tintColor: .blue, duration: "120 mins", date: "Aug 11", calories: "884 kcal")
-    ]
-    
+
     var body: some View {
         
         NavigationStack {
@@ -42,14 +27,14 @@ struct HomeView: View {
                         
                         Spacer()
                         
-                        VStack {
+                        VStack(alignment: .leading) {
                             VStack(alignment: .leading, spacing: 8) {
                                 Text("Calories")
                                     .font(.callout)
                                     .bold()
                                     .foregroundStyle(.red)
                                 
-                                Text("123 kcal")
+                                Text("\(viewModel.calories) kcal")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -60,7 +45,7 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundStyle(.green)
                                 
-                                Text("52 mins")
+                                Text("\(viewModel.active) mins")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -71,7 +56,7 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundStyle(.blue)
                                 
-                                Text("8 hours")
+                                Text("\(viewModel.stand) h")
                                     .bold()
                             }
                         }
@@ -79,12 +64,12 @@ struct HomeView: View {
                         Spacer()
                         
                         ZStack {
-                            ProgressCircleView(progress: $calories, color: .red, goal: 600)
+                            ProgressCircleView(progress: $viewModel.calories, color: .red, goal: 600)
                             
-                            ProgressCircleView(progress: $active, color: .green, goal: 60)
+                            ProgressCircleView(progress: $viewModel.active, color: .green, goal: 60)
                                 .padding(.all, 20)
                             
-                            ProgressCircleView(progress: $stand, color: .blue, goal: 12)
+                            ProgressCircleView(progress: $viewModel.stand, color: .blue, goal: 12)
                                 .padding(.all, 40)
                         }
                         .padding(.horizontal)
@@ -114,7 +99,7 @@ struct HomeView: View {
                     
                     LazyVGrid(columns: Array(repeating: GridItem(spacing: 29),
                                              count: 2)) {
-                        ForEach(mockActivities, id: \.id) { activity in
+                        ForEach(viewModel.mockActivities, id: \.id) { activity in
                             ActivityCardView(activity: activity)
                         }
                     }
@@ -141,7 +126,7 @@ struct HomeView: View {
                     .padding(.top)
                     
                     LazyVStack {
-                        ForEach(mockWorkouts, id: \.id) { workout in
+                        ForEach(viewModel.mockWorkouts, id: \.id) { workout in
                             WorkoutCardView(workout: workout)
                         }
                     }
