+++ FitnessApp/Home/Views/WorkoutCardView.swift
@@ -8,16 +8,6 @@
 import SwiftUI
 
 
-struct Workout {
-    let id: Int
-    let title: String
-    let image: String
-    let tintColor: Color
-    let duration: String
-    let date: String
-    let calories: String
-}
-
 struct WorkoutCardView: View {
     
     @State var workout: Workout
